+++ includes/config.kate.php
@@ -3,13 +3,20 @@
 // Project Site www.aas9.in/zerocms
 // Created March 2014
 
-// if on a shared server
+// enter site URL/domain here
 $site = 'http://mywebsite.com'; //without a trailing slash
-$host = 'localhost/projects/zerocms';
 
 // if on a private server use this
 /* $protocol = 'http://';
 $host = gethostname();
 $site = $protocol . $host;
 */
+?>
+<?php
+//enter site name
+$site_name = 'ZeroCMS';
+$title = $site_name;
+
+//enter disqus comments shortname (more info here- https://help.disqus.com/customer/portal/articles/466208-what-s-a-shortname-)
+$disqus_comments_name = 'zerocms';
 ?>