+++ includes/db.kate.php
@@ -4,6 +4,6 @@
 // Created March 2014
 define('MYSQL_HOST', 'localhost');
 define('MYSQL_USER', 'root');
-define('MYSQL_PASSWORD', '');
+define('MYSQL_PASSWORD', 'password');
 define('MYSQL_DB', 'zerocms');
 ?>