+++ views/zero_login.php
@@ -2,26 +2,40 @@
 // (c)Perez Karjee(www.aas9.in)
 // Project Site www.aas9.in/zerocms
 // Created March 2014
+require_once '../includes/config.kate.php';
 include '../includes/header.kate.php';
+include '../includes/wrapper-start.php';
+include '../includes/menu.kate.php';
 ?>
-<h1>Member Login</h1>
-<form method="post" action="zero_transact_user.php">
-<table>
-	<tr>
-		<td><label for email="email">Email Address:</label></td>
-		<td><input type="text" id="email" name="email" maxlength="100"/></td>
-	</tr>
-	<tr>
-		<td><label for="password">Password</label></td>
-		<td><input type="password" id="password" name="password" maxlength="20"/></td>
-	</tr>
-	<tr>
-		<td> </td>
-		<td><input type="submit" name="action" value="Login"/></td>
-		</tr>
-</table>
-</form>
-<p>Not A Member Yet ? <a href="zero_user_account.php">Create A New Account</a></p>
-<p><a href="zero_forgot_password.php">Forgot Your Password ?</a></p>
-<?php include '../includes/footer.kate.php';
+
+<div class="content_bottom">
+<div class="grid_1_of_2 box">
+
+<?php
+
+if(isset($_SESSION['user_id'])){ //if logged in
+echo '<h3 align="center">Already Logged In</h3><meta http-equiv="refresh" content="1;url=zero_cpanel.php" />';
+} else{
+
+		echo '<h3>Login</h3>
+				<div class="searh_form">
+					<form method="post" action="zero_transact_user.php">
+						<input type="text" id="email" name="email" placeholder="hello@yourmail.com" class="active" maxlength="120"/>
+						<input type="password" id="password" name="password" placeholder="password" class="active" >
+						<input type="submit" name="action" value="Login" />
+					</form>
+				</div>
+				<br>
+				<p>Not A Member Yet ? <a href="zero_user_account.php">Create A New Account</a></p>
+				<p><a href="zero_forgot_password.php">Forgot Your Password ?</a></p>
+			</div>
+	</div>';
+		}
+
+?>
+
+<?php
+//end
+include '../includes/wrapper-end.php';
+include '../includes/footer.kate.php';
 ?>