+++ changelog.md
@@ -10,3 +10,15 @@ ChangeLog
 	version 1.2 released [stable]
 	date- 30 oct 2014
 	changes- lots of bug fixes and some small changes too
+	
+	version 1.3 released [beta]
+	date- 3 nov 2014
+	changes- change in architecture, a complete new look, lot of bug fixes & security patches
+	
+	version 1.3.1 released [RC]
+	date- 4 nov 2014
+	changes- bug fixes and small changes
+	
+	version 1.3.2 released [RC]
+	date- 4 nov 2014
+	changes- bug fixes, small changes & disqus comments integration